Update perceptron learning rate and readme file Merge branch 'master' of https://github.com/ACHYLAND/cse353proj
# Conflicts:
#	cse353/matrixManipulation.java
#	cse353/test.java
#	linearRegression.java

+++ cse353/CSVReader.java
@@ -7,19 +7,19 @@ import java.util.Vector;
 
 class CSVReader {
 
-    ArrayList<double[]> read(String filename) {
+    ArrayList<int[]> read(String filename) {
             BufferedReader br;
             String line;
             final String COMMA_DELIMITER = ",";
 
             try {
                 br = new BufferedReader(new FileReader(filename));
-                ArrayList<double[]> vectors = new ArrayList<>();
+                ArrayList<int[]> vectors = new ArrayList<>();
 
                 int j = 0;
                 while ((line = br.readLine()) != null) {
                     String[] fvs = line.split(COMMA_DELIMITER);
-                    double[] featureVector = new double[fvs.length];
+                    int[] featureVector = new int[fvs.length];
 
                     for (int i = 0; i < fvs.length; i++) {
                         featureVector[i] = Integer.parseInt(fvs[i]);
